% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray.R
\name{wkxray_command}
\alias{wkxray_command}
\title{Apply a command JSON and return the resulting response JSON.}
\usage{
wkxray_command(xray, cmd_json)
}
\arguments{
\item{xray}{An xray handle from [wkxray_new()].}

\item{cmd_json}{A command JSON string.}
}
\value{
The response as a JSON string.
}
\description{
Apply a command JSON and return the resulting response JSON.
}
